UI Improvements + Misc Fixes
- Removed broken Twitter API support
- Switched from Fontawesome to IonIcons
- Formatting
-  Changed to generic account type (user/password)
- Added login validation
- Updated NodeJS
- Script improvements
- Removed topbar (UI)
-  Increased period access token is valid for
