+++ .env.example
@@ -13,5 +13,3 @@ SSL_KEY='/ssl/key.pem'
 SSL_CRT='/ssl/cert.pem'
 API_CORS_ALLOWED='*'
 LOG_LEVEL='DEBUG'
-TWITTER_CLIENT_ID=''
-TWITTER_CLIENT_SECRET=''
