+++ api/app/application/post.js
@@ -1,4 +1,4 @@
-import { post as twitterPost } from "./twitter.js";
+// import { post as twitterPost } from "./twitter.js";
 import { schedule, unschedule } from "./schedule.js";
 import auth from "./authentication.js";
 const { authenticateToken } = auth;
