+++ api/app/models/account.js
@@ -1,0 +1,16 @@
+import pkg from "mongoose";
+const { Schema } = pkg;
+
+export const accountSchema = new Schema(
+  {
+    userEmail: {
+      type: String,
+      required: true,
+    },
+    password: {
+      type: String,
+      required: true,
+    },
+  },
+  { timestamps: true }
+);
