+++ dev-compose.yaml
@@ -44,8 +44,6 @@ services:
       API_PORT: ${API_PORT}
       SECRET_1: ${SECRET_1}
       SECRET_2: ${SECRET_2}
-      TWITTER_CLIENT_ID: ${TWITTER_CLIENT_ID}
-      TWITTER_CLIENT_SECRET: ${TWITTER_CLIENT_SECRET}
       TIMEZONE_OFFSET: ${TIMEZONE_OFFSET}
       LOG_LEVEL: ${LOG_LEVEL}
     volumes:
