--- web/www/assets/js/navigation.js
@@ -1,6 +1,0 @@
-export function hasMobileMenu() {
-  var w = parseInt(screen.width);
-  if (w <= 750) {
-    //mobile menu code here
-  }
-}
