+++ web/www/assets/package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "ionicons": "^7.1.2"
+  }
+}
