+++ web/www/html/header.html
@@ -11,7 +11,8 @@
       href="https://fonts.googleapis.com/css2?family=Poppins:wght@100;200;400;600;700&display=swap"
       rel="stylesheet"
     />
-    <link href="assets/fontawesome/css/all.css" rel="stylesheet" />
-    <link rel ="icon" href ="assets/img/icon.ico" type = "image/x-icon">
+    <link rel="icon" href="assets/img/icon.ico" type="image/x-icon" />
+    <script type="module" src="assets/node_modules/ionicons/dist/ionicons/ionicons.esm.js"></script>
+    <script nomodule src="assets/node_modules/ionicons/dist/ionicons/ionicons.js"></script>
   </head>
 </html>
