+++ web/www/html/modal-popup.html
@@ -1,7 +1,7 @@
 <div class="bg-modal" id="bgmodal">
   <div class="modal-content" id="modalContent">
     <div class="close" id="closePopup">+</div>
-    <h3 id="infoHeading"></h3>
+    <h3 id="infoHeading" class="info-heading"></h3>
     <div id="infoDescription"></div>
   </div>
 </div>
