--- web/www/html/topbar.html
@@ -1,13 +1,0 @@
-<div class="topbar">
-  <div class="search">
-    <label>
-      <input type="text" placeholder="search" />
-      <i class="fas fa-search"></i>
-    </label>
-  </div>
-  <div class="topbar-markers" id="topbarMarkers"></div>
-  <div class="userimg">
-    <!-- <img src="../assets/img/user.png" alt="user image"> -->
-    <i class="fa-solid fa-user-astronaut"></i>
-  </div>
-</div>
