+++ web/www/index.php
@@ -1,19 +1,18 @@
 <?php
-  session_start();
-  session_regenerate_id();
-  if( !isset($_SESSION['token']) )
-  {
-    header("Location: /login");
-  }
+session_start();
+session_regenerate_id();
+if (!isset($_SESSION['token'])) {
+  header("Location: /login");
+}
 ?>
-<?php include(__DIR__ .  '/html/header.html') ?>
+<?php include(__DIR__ . '/html/header.html') ?>
+
 <body>
   <section class="body-section">
     <?php include(__DIR__ . '/html/sidebar.html') ?>
     <div class="main">
-      <?php include(__DIR__ . '/html/topbar.html') ?>
       <div id="app" class="app">
-          <div id="appScreen" class="app-screen"></div>
+        <div id="appScreen" class="app-screen"></div>
       </div>
     </div>
     <?php include(__DIR__ . '/html/modal-popup.html') ?>
@@ -23,6 +22,5 @@
 
   <script src="assets/js/modules/masonry/masonry.pkgd.min.js"></script>
   <script src="assets/js/modules/masonry/imagesloaded.pkgd.js"></script>
-  <script src="assets/js/app.js" type="module" ></script>
+  <script src="assets/js/index.js" type="module"></script>
 </body>
-<?php include(__DIR__ . '/html/footer.html') ?>
